- added reactotron for logs monitoring - added loggerutil - update readme - added new sripts - install reactotron plugins
