+++ index.js
@@ -1,7 +1,7 @@
 /**
  * @format
  */
-
+import './reactotron.config'
 import {AppRegistry} from 'react-native';
 import App from './App';
 import {name as appName} from './app.json';
