+++ package.json
@@ -7,9 +7,24 @@
     "ios": "react-native run-ios",
     "start": "react-native start",
     "test": "jest",
-    "lint": "eslint . --ext .js,.jsx,.ts,.tsx"
+    "lint": "eslint . --ext .js,.jsx,.ts,.tsx",
+    "android-hermes": "npx react-native run-android --variant release",
+    "android-stage": "set ENVFILE=.env.staging && npx react-native run-android",
+    "android-prod": "set ENVFILE=.env.production && npx react-native run-android",
+    "aab": "cd android && ./gradlew bundleRelease",
+    "apk": "cd android && ./gradlew assembleRelease",
+    "apk-staging": "cd android && ENVFILE=.env.staging ./gradlew assembleRelease",
+    "apk-prod": "cd android && ENVFILE=.env.production ./gradlew assembleRelease",
+    "pods": "cd ios && pod install && cd ..",
+    "adb-start": "adb reverse tcp:9090 tcp:9090 && node node_modules/react-native/local-cli/cli.js start",
+    "clean": "cd android && ./gradlew clean",
+    "reset": "yarn start - --reset-cache",
+    "tcp": "adb reverse tcp:8081 tcp:8081",
+    "both": "react-native run-android && react-native run-ios",
+    "deintegrate": "cd ios && pod deintegrate && pod update && cd .."
   },
   "dependencies": {
+    "@react-native-async-storage/async-storage": "^1.15.17",
     "react": "17.0.2",
     "react-native": "0.66.4"
   },
@@ -18,6 +33,7 @@
     "@babel/runtime": "^7.12.5",
     "@react-native-community/eslint-config": "^2.0.0",
     "@types/jest": "^26.0.23",
+    "@types/lodash": "^4.14.178",
     "@types/react-native": "^0.66.4",
     "@types/react-test-renderer": "^17.0.1",
     "@typescript-eslint/eslint-plugin": "^5.7.0",
@@ -27,6 +43,9 @@
     "jest": "^26.6.3",
     "metro-react-native-babel-preset": "^0.66.2",
     "react-test-renderer": "17.0.2",
+    "reactotron-apisauce": "^3.0.0",
+    "reactotron-react-native": "^5.0.1",
+    "reactotron-redux": "^3.1.3",
     "typescript": "^4.4.4"
   },
   "resolutions": {
